restructured, extracted classes to separate files
